Improved the Measure class
Now it is possible to give a custom integration for each measure. It is also possible to restrict the computation of a given measure to the set of parameters for which the PDF is greater than a given threshold, reducing the computational cost at the price of a (small) approximation.